add elapsed time metrics in api serve framework
